start working on hamiltonian terms
