+++ lib/EaRydHamiltonians/src/EaRydHamiltonians.jl
@@ -1,0 +1,7 @@
+module EaRydHamiltonians
+
+using Configurations
+
+include("types.jl")
+
+end # module
