+++ lib/EaRydHamiltonians/src/show/show.jl
@@ -1,0 +1,1 @@
+include("plain.jl")
